+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ## [ADD NEW VERSION HERE] - ADD DATE HERE
 ### Fixed
 ### Added
+ - Add '--hwp-min-perf' and '--hwp-max-perf' options for reading and setting
+   HWP min/max performance levels via MSR.
 ### Removed
 ### Changed